requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
python-dotenv==0.21.0
python-json-logger==2.0.4
requests==2.28.1
sendgrid==6.9.7
